+++ packages/hightide-native/src/theme/resolvers/chip.ts
@@ -1,6 +1,8 @@
 import { toContainerStyle } from '../adapters/container-adapter'
+import { toIconStyle } from '../adapters/icon-style-adapter'
 import { toTextStyle } from '../adapters/text-style-adapter'
 import type {
+  ChipIconStyle,
   ChipState,
   ChipStyle,
   ChipTextStyle,
@@ -8,6 +10,7 @@ import type {
 } from '../types/components/chip'
 import {
   createStyleResolver,
+  createValueResolver,
   type ComponentThemeResolver
 } from '../types/resolver'
 
@@ -29,6 +32,9 @@ export const toChipThemeResolvers: ComponentThemeResolver<ChipThemeResolvers> = 
   return {
     chip: createStyleResolver((state: ChipState): ChipStyle => (
       toContainerStyle(resolve(state).container)
+    )),
+    icon: createValueResolver((state: ChipState): ChipIconStyle => (
+      toIconStyle(resolve(state).icon)
     )),
     text: createStyleResolver((state: ChipState): ChipTextStyle => (
       toTextStyle(resolve(state).text)