+++ packages/hightide-native/src/theme/types/components/chip.ts
@@ -7,6 +7,7 @@ import type { ColorPairToken } from '@helpwave/hightide-design/theme-tokens'
 import type { ChipVariant, ComponentSize } from '@helpwave/hightide-design/semantic-token-resolvers'
 
 import type { StyleResolverFunction } from '../resolver'
+import type { IconStyle } from '../../../icons'
 
 export type ChipState = {
   size?: ComponentSize,
@@ -18,7 +19,10 @@ export type ChipStyle = ViewStyle
 
 export type ChipTextStyle = TextStyle
 
+export type ChipIconStyle = IconStyle
+
 export type ChipThemeResolvers = {
   chip: StyleResolverFunction<ChipState, ChipStyle>,
+  icon: StyleResolverFunction<ChipState, ChipIconStyle>,
   text: StyleResolverFunction<ChipState, ChipTextStyle>,
 }